requirements.txt
flask
Flask-WTF
WTForms
maeapi
peewee
python-dateutil